refactor(Cavalo): swith back from Item to Cavalo
