+++ front-end/src/css/Cavalo.css
@@ -1,27 +1,27 @@
 /* Item.css */
 
-.item {
+.cavalo {
     display: flex;
     align-items: center;
     padding: 10px 0;
   }
   
-  .item-icone {
+  .cavalo-icone {
     color: #1abc9c;
     margin-right: 15px;
   }
   
-  .item-detalhes {
+  .cavalo-detalhes {
     flex-grow: 1;
   }
   
-  .item-nome {
+  .cavalo-nome {
     font-size: 22px;
     color: #2c3e50;
     margin: 0;
   }
   
-  .item-info {
+  .cavalo-info {
     font-size: 16px;
     color: #7f8c8d;
     margin: 5px 0 0;
